Ensure `oauth_client_class` sends `Authorization` HTTP header.
Apparently, Manuel Lemos's `oauth_client_class` will not send the
`Authorization` HTTP header if the `access_token_type` is not set
explicitly, but the `access_token` is set explicitly on a `GET` request.
This is true even if the `authorization_header` variable is set
explicitly. In order to make sure that even `GET` requests have the
`Authorization` header added, I must explicitly set the
`access_token_type` as well as the `access_token` variables when using
OAuth 2.0 connections.

git-svn-id: https://plugins.svn.wordpress.org/wp-crosspost/trunk@884466 b8457f37-d9ea-0310-8a92-e5e31aec5664